+++ lipoic/cogs/fun/chef.py
