--- locales/en.json
@@ -1,3 +1,0 @@
-{
-  "test": "awa"
-}
